Added unit testing, fixed file structure.
