+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./styles.css">
-    <script src="./script.js" type="module" defer></script>
+    <script src="./scripts/script.js" type="module" defer></script>
     <title>Morse Code Translator</title>
 </head>
 <body>
